BAVL-1251 refactoring in preparation for reschedule booking emails work, no functional changes.
